# Performance updates for virtual_attribute_search