Fixes two simple performance issues:

- Avoids trying to Rbac records that don't need to be "Rbac'd" (first commit)
- Avoids repeating Rbac on records that have already been checked in the initial `collection_search` (second commit)


Benchmarks
----------

**Before**

```
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
| 2197 |     322 |      194.4 | 1627 |
|  953 |     321 |      182.7 |  196 |
|  996 |     321 |      187.1 |  196 |
```

**After** (first commit)

```
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
| 2274 |     282 |      166.2 | 1612 |
|  858 |     281 |      147.9 |  181 |
|  960 |     281 |      164.2 |  181 |
```

**After** (second commit)

```
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
| 2004 |     263 |      156.8 | 1593 |
|  756 |     261 |      142.9 |  161 |
|  769 |     261 |      138.9 |  161 |
```


TODO
----

* [x] Make sure existing tests pass...
* [x] Add Tests


Links
-----

* Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1656371
* Related to changes in https://github.com/ManageIQ/manageiq/pull/15145