@NickLaMuro 2019-02-08
@lpichler Can I have you take a look at this and make sure this still falls in line with the goals of https://github.com/ManageIQ/manageiq/pull/15145 ?  I might adds some tests before calling this good, but would at least like your opinion if what I changed here seems kosher.  Thanks!