@lpichler 2019-02-14
> @lpichler Can I have you take a look at this and make sure this still falls in line with the goals of [ManageIQ/manageiq#15145](https://github.com/ManageIQ/manageiq/pull/15145) ? I might adds some tests before calling this good, but would at least like your opinion if what I changed here seems kosher. Thanks!

@NickLaMuro Looks good to me, I tested out - nice finding - just question to - would be helpful to  bring your `Better Rbac-able check` mechanism from first commit also for [this RBAC call ](https://github.com/ManageIQ/manageiq-api/pull/557/files#diff-00bf332c6d31e69fa897ff19cd3e0bc5R211)?

thanks!