@NickLaMuro 2019-02-14
@lpichler Thanks for taking a look!

> would be helpful to bring your `Better Rbac-able check` mechanism from first commit also for [this RBAC call ](https://github.com/ManageIQ/manageiq-api/pull/557/files#diff-00bf332c6d31e69fa897ff19cd3e0bc5R211)?

Oh, I assume you mean this part:

```ruby
return resource_attr unless Rbac::Filterer.new.send(:apply_rbac_directly?, klass)
```

Then yeah, I think this makes sense to do.  I can add that and hopefully getting to writing some tests for this today as well.

Thanks again!