@kbrock 2019-02-26
Note: This is for speeding up https://bugzilla.redhat.com/show_bug.cgi?id=1656371