@kbrock 2019-03-21
@NickLaMuro does it make sense to add this bail shortcut to Rbac.filtered?