@NickLaMuro 2019-03-25
Added some specs on this one, so removed the specs on this as I feel a little more confident now with some tests validating the changes.